Added check for register_argc_argv
Adds a hint to an issue that had me clueless for awhile.

I have a strange scenario where my shell user is different to my cPanel Cron Manager and run from two different configurations. This lead to a vicious circle of wondering why it works in SSH, but not within the cron manager.