Remove module-specific docs instructions
We have too many docs to list these out now,
even in 3.x this was a bit of a stopgap solution.
Point to a centrally managed URL on silverstripe.org
instead, where we can update the list of "core modules" regularly
without breaking URLs in the docs etc

Note that these URLs are also used internally by the
Open Sourcerers team.