Fixing HTMLEditorField API documentation
The SS3 API documentation for HTMLEditorField was broken before change #7904:
http://api.silverstripe.org/3/HtmlEditorField.html

This was because there was an unescaped textarea tag in the class description. In #7904 we escaped the tag. 

However in a similar pull request #7927 it was suggested we should use backticks instead. 

This changes the documentation to use backticks around the textarea tag.