Update jquery-ui.min.js to fix minified syntax bug
In two places in this file there is the syntax `+ ++` which throws an interpreter error and causes all javascripts in the same file (such as when combining this file with other files) to fail to load. This problem can easily be solved by wrapping the ++n and ++t with parentheses.
This bug is discussed at:
https://bugs.jqueryui.com/ticket/5576
and 
http://www.crockford.com/javascript/jsmin.html

I'm not sure why nobody else has ever encountered this issue. Till now I've been manually editing this one file after any framework updates.