FIX allow extensions to be applied via injector service definitions
Throughout the code there is the implication of being able to apply
extensions through an Injector service name - such as the stripping
of %$ before looking up an extensions name.

However this is not possible as class_exists is exclusively used to
ensure that the Extension is a valid subclass of Extension (or
DataExtension).

This commit seeks to rectify that by falling back to using Injector
to see if there is a service registered with the given name, and if
so using it via instantion before performing the sanity check of
the concrete class' inheritance heirarchy.