FIX allow for Injector service dot notation to fall-back to classes
A feature was introduced to Injector with SilverStripe 4 allowing
developers to define services in the format
ServiceName.SubAlias
allowing for ServiceName to have contextual information to allow it
to be used in a modified way for a particular situation where that
context is relevant. The idea being that if a particular alias of
the service proper was undefined, instantiation will fall back to
the service proper.

Injector also allows for an undefined service to be treated as a
classname, and attempt to instantiate it as such.

However the two features are incompatible, a service will only fall
back to another pre-defined service, which is contrary to expectation
if the fallback is a valid classname, and an undefined service.

So we can fix this by treating the section of the service defintion
preceeding a dot as we would with any service definition without a
dot - that is if the service is undefined, to treat the name as a
fully qualified class name.