FIX Filter hostname in TrustedProxyMiddleware
In depth defense, rather than a security vulnerability.
The trusted proxy IP list is empty by default, meaning no request can set this information.
Systems need to opt in to this behaviour by setting specific proxy IPs.
While there is an "all IPs" option, it's not documented on
https://docs.silverstripe.org/en/4/developer_guides/security/secure_coding/#request-hostname-forgery.
If you find and configure this option, you're also expected to sanitise these layers (e.g. a load balancer).

This issue was originally referred to as SS-2018-024