FIX enable TinyMCE anchor plugin
The sslink anchor module (which is enabled by default) allows an editor to insert links to named anchor elements contained in text on other pages in SilverStripe CMS. However there was no way to add an anchor besides the complicated task of directly editing HTML via TinyMCE's plugin for that. This was existing functionality in SilverStripe 3, but seems to have been lost with 4.0.0 onwards.