Adding SS_ENVIRONMENT_TYPE to documentation
Before running a dev/build the first time, you need to specify dev as your environment type. The variable wasn't mentioned as part of the list so I've added it. Let me know if it's not clear about the difference between the states, or it should be better documented here somehow.