FIX Mitigate continue; warning when run on PHP 7.3+ (fixes #9826)
PHP 7.3+ issues a warning when continue; is used inside a switch statement, as it has the same behaviour as break; HTMLBBCodeParser.php had such a continue; statement that is intended to skip to the the next record of the foreach that the switch sits inside. continue 2; needs to be used here.