DOC: fix broken heading
It appears there may be a bug in gatsby which causes code blocks in headings to not be output. I can't find a specific issue in their project to link this to, so it might be best to simply remove the code block here allowing the header to display correctly.