CanonicalURLMiddleware default to all environments
Having a Live/Prod only feature sets people up for bugs. 
A common scenario for this one is a javascript feature that requests from an API - its easy to forget a trailing slash in this scenario and it will work on your local an your test environment - failing only once you deploy it to prod. I'd suggest we turn this on everywhere except the CLI by default (or, disable it by default).