FIX 'which' check in 'sake' now works for aliases
Improves on the current implementations since
'command' already only returns executable commands.
Additionally, the success output of 'command' is not standardized
and may contain additional information.
Therefore, a 'test' if the result of 'command'
is a valid executable file, may fail incorrectly.

Now just relying on the standardized exit status of 'command -v'.