FIX: GRIDFIELD - fixing error interaction between limit and itemsPerPage
If the total list has less than the items per page you need to make sure that you set the items per page to be the total number of items otherwise, you still get for example 20 items, in the gridfield, but the paginator info will say 1 of 1. Basically, the paginator override the limit, but should not do so if the limit is less than the items per page.