ENH Refactor password reset to support SameSite=Strict
Performing an immediate redirect on a request from an external website,
such as a web-based email client, causes the new request to be treated
as external, and when the session cookie is set to Samesite=Strict, this
prevents the cookie from being sent by the browser, triggering a fresh
session. This meant that the existing password reset mechanism would
not function in this mode, as the AutoLoginHash was being stored in the
session and immediately lost, triggering a redirect to the login form.

This change refactors the change password handler to instead push the
AutoLoginHash value into the change password form as a hidden field,
ensuring it can be read during submission.

It also includes broader test coverage of the change password handler,
though this remains incomplete due to time constraints.