MNT Don't fail test for unrelated line number change
symfony/yaml changed which line it thinks the error is on - arguably the
change is wrong, but arguably the line number was already wrong anyway
(the actual error is on line 4, but we used to expect line 5, and it
changed to reporting line 6).

We shouldn't care what line symfony is reporting about - what really
matters is that the error we expect is the error that is thrown.