FIX Make SearchableDropdownTrait::dataValue() return correct value
This method is meant to return a value that's ready for inserting into
the database - for these form fields that means the raw ID value or
array of ID values.

Previously this returned the raw submitted value, which was an associative
array that included the value, the label, and whether the value was
selected or not - which is not useful.